# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Logenhelfer API
  version: 1.0.0
  description: Backend-agnostic contract. Rails serves it today, Express later.
paths:
...
  /api/v1/lodges:
    get:
      operationId: listLodges
      parameters:
        - name: page
          in: query
          schema: { type: integer, minimum: 0, default: 0 }
        - name: per_page
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
        - name: sort
          in: query
          schema: { type: string, default: name }
      responses:
        '200':
          description: Paginated lodge list
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LodgeList' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      operationId: createLodge
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LodgeInput' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Lodge' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /api/v1/lodges/{slug}:
    parameters:
      - name: slug
        in: path
        required: true
        schema: { type: string }
    get:
      operationId: getLodge
      responses:
        '200':
          description: Lodge detail
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Lodge' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      operationId: updateLodge
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LodgeInput' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Lodge' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
    delete:
      operationId: deleteLodge
      responses:
        '204': { description: Soft-deleted (cascades to its officers) }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  responses:
    Unauthorized:
      description: Missing or invalid credentials
      content:
        application/json:
          schema:
            type: object
            required: [error]
            properties:
              error: { type: string }
...
    Forbidden:
      description: Authenticated but not permitted
      content:
        application/json:
          schema:
            type: object
            required: [error]
            properties:
              error: { type: string }
...
    NotFound:
      description: No record with that identifier
      content:
        application/json:
          schema:
            type: object
            required: [error]
            properties:
              error: { type: string }
    UnprocessableEntity:
      description: Validation failed
      content:
        application/json:
          schema:
            type: object
            required: [error, detail]
            properties:
              error: { type: string }
              detail: { type: string }
  schemas:
...
    LodgeSummary:
      type: object
      additionalProperties: false
      required: [slug, name, description, district_name]
      properties:
        slug: { type: string }
        name: { type: string }
        description: { type: string, nullable: true }
        district_name: { type: string }
    Lodge:
      type: object
      additionalProperties: false
      required: [slug, name, description, district_id, district_name]
      properties:
        slug: { type: string }
        name: { type: string }
        description: { type: string, nullable: true }
        district_id: { type: integer }
        district_name: { type: string }
    LodgeList:
      type: object
      additionalProperties: false
      required: [rows, row_count]
      properties:
        rows:
          type: array
          items: { $ref: '#/components/schemas/LodgeSummary' }
        row_count: { type: integer }
    LodgeInput:
      # No top-level `required` - same reasoning as CategoryInput/
      # SeekerInput: the model's own presence validations (name,
      # district_id) produce the 422 detail message. district_id is
      # accepted on BOTH create and update - unlike Directory's
      # category_slug, Lodge's district assignment is NOT immutable
      # after creation (see Global Constraints).
      type: object
      additionalProperties: false
      properties:
        name: { type: string }
        description: { type: string, nullable: true }
        district_id: { type: integer }
...
security:
  - bearerAuth: []